Add input/move, asteroids, and sprite refactor
Introduce MoveComponent and InputComponent plus Random utility; add Asteroid actor and wire ship to use Input/Move for player control. Refactor SpriteComponent/TileMap to use rotation, center, flip modes, and scale sprites by actor scale. Propagate SDL events from Game -> Actor -> Component, update Actor/Component headers, and change window dimensions to floats. Remove old Texture abstraction and related files, update CMakeLists and .gitignore, and adjust various game loading/rendering bits to support asteroids and the new input/movement system.

+++ Engine/include/Constants.h
@@ -8,8 +8,8 @@
 #include <string>
 
 inline std::string windowTitle = "Game Programming with C++";
-inline int windowWidth{ 1920 };
-inline int windowHeight{ 1080 };
+inline float windowWidth{ 1920 };
+inline float windowHeight{ 1080 };
 inline int screenFPS{ 60 };
 
 inline bool vSyncEnabled{ true };
